pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "sana-aortic-dissection"
authors = [
    {name = "SaNa Research Team", email = "research@sana-project.org"},
]
description = "Spec-Kit for Aortic Dissection Analysis - Multi-center AI prediction system"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["medical-ai", "aortic-dissection", "machine-learning", "multimodal-fusion"]
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "scipy>=1.7.0",
    "torch>=1.9.0",
    "torchvision>=0.10.0",
    "lightgbm>=3.3.0",
    "xgboost>=1.5.0",
    "SimpleITK>=2.1.0",
    "pyradiomics>=3.0.0",
    "scikit-image>=0.19.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
    "statsmodels>=0.13.0",
    "PyYAML>=6.0",
    "tqdm>=4.62.0",
    "openpyxl>=3.0.0",
    "xlrd>=2.0.0",
    "tables>=3.6.0",
    "h5py>=3.6.0",
    "shap>=0.41.0",
    "joblib>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.2.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
    "pre-commit>=2.17.0",
]
docs = [
    "sphinx>=4.4.0",
    "sphinx-rtd-theme>=1.0.0",
    "myst-parser>=0.18.0",
]
tracking = [
    "wandb>=0.12.0",
    "mlflow>=1.20.0",
]

[project.urls]
Homepage = "https://github.com/sana-project/sana-aortic-dissection"
Repository = "https://github.com/sana-project/sana-aortic-dissection"
Documentation = "https://sana-aortic-dissection.readthedocs.io/"
"Bug Tracker" = "https://github.com/sana-project/sana-aortic-dissection/issues"

[project.scripts]
sana-etl = "etl.cli:main"
sana-train = "models.cli:main"
sana-eval = "eval.cli:main"

[tool.setuptools]
packages = ["etl", "features", "models", "eval", "tests"]

[tool.setuptools.package-data]
etl = ["../configs/*.yml"]
configs = ["*.yml"]
info = ["*.txt", "*.pdf"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["etl", "features", "models", "eval"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]